fix: Fix logs page pagination and button cursor styles (v0.44.3)
Two fixes for the logs page:

1. Fixed pagination sequence function type mismatch
   - Changed sequence template function from int64 to int
   - Template was calling sequence(1, $totalPages) with int values
   - Function signature was expecting int64, causing type error
   - Error: "wrong type for value; expected int64; got int"
   - Pagination is limited to int range anyway, int64 unnecessary

2. Added cursor-pointer to filter action buttons
   - Apply Filters, Clear, and Refresh buttons now show pointer cursor
   - Improves UX by indicating buttons are clickable

Changes:
- internal/server/handlers.go: sequence function int64 → int
- web/templates/logs.html: Added cursor-pointer class to 3 buttons

+++ internal/server/handlers.go
@@ -3682,14 +3682,14 @@ func (s *Server) createTemplateFuncs() template.FuncMap {
 			}
 			return result
 		},
-		"sequence": func(start, end int64) []int64 {
+		"sequence": func(start, end int) []int {
 			// Generate a sequence of integers from start to end (inclusive)
 			if start > end {
-				return []int64{}
+				return []int{}
 			}
-			result := make([]int64, end-start+1)
+			result := make([]int, end-start+1)
 			for i := range result {
-				result[i] = start + int64(i)
+				result[i] = start + i
 			}
 			return result
 		},
